requirements.txt
python-telegram-bot==20.3
google-genai==0.1.0
python-dotenv==1.1.1
websockets==14.2
httpx==0.24.1
httpcore==0.17.3
pydantic==2.11.9
pydantic-core==2.33.2
requests==2.32.5
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
h11==0.14.0
anyio==4.10.0
typing-extensions==4.15.0
